解决Tux模式下子弹与人物无碰撞的问题

+++ assets/src/game/map/object/iwbt/boss/BossBullet.ts
@@ -1,8 +1,11 @@
-import { _decorator } from "cc";
+import { Collider2D, Contact2DType, IPhysics2DContact, _decorator } from "cc";
 import { noxcc } from "../../../../../framework/core/noxcc";
+import { GameConfig } from "../../../../config/GameConfig";
+import { ObjectGroup } from "../../../../const/ObjectGroup";
+import { ObjectTag } from "../../../../const/ObjectTag";
 import { CameraControl } from "../../../CameraControl";
-import { MapUtil } from "../../../MapUtil";
 import { BaseObject } from "../../BaseObject";
+import { Player } from "../../Player";
 
 const { ccclass, property, executeInEditMode, disallowMultiple, requireComponent, executionOrder } = _decorator;
 
@@ -12,6 +15,12 @@ export class BossBullet extends BaseObject {
     public speedX: number = 0;
     public speedY: number = 0;
 
+    public autoRemove: boolean = false;
+
+    start() {
+        this.getComponent(Collider2D).on(Contact2DType.BEGIN_CONTACT, this.onBeginContact, this);
+    }
+    
     update(dt: number): void {
         if (this.speedX || this.speedY) {
             noxcc.addXY(this.node, this.speedX * dt, this.speedY * dt);
@@ -20,17 +29,30 @@ export class BossBullet extends BaseObject {
         var pos = noxcc.convertPosAR(noxcc.pos(this.node), this.node.parent, this.map.node);
         pos.x += noxcc.aw(this.map.node);
         pos.y += noxcc.ah(this.map.node);
-        var cameraControl = this.map.getComponent(CameraControl);
-        var cameraRect = cameraControl.getCameraRect();
-        cameraRect.x -= 150;
-        cameraRect.y -= 150;
-        cameraRect.width += 300;
-        cameraRect.height += 300;
-        if (!cameraRect.contains(pos)) {
-            this.node._destroyImmediate();
+
+        if(this.autoRemove) {
+            var cameraControl = this.map.getComponent(CameraControl);
+            var cameraRect = cameraControl.getCameraRect();
+            cameraRect.x -= 150;
+            cameraRect.y -= 150;
+            cameraRect.width += 300;
+            cameraRect.height += 300;
+            if (!cameraRect.contains(pos)) {
+                this.node._destroyImmediate();
+            }
         }
     }
 
+    private onBeginContact(selfCollider: Collider2D, otherCollider: Collider2D, contact: IPhysics2DContact): void {
+        if (otherCollider.tag != ObjectTag.Default) return;
+        if (ObjectGroup.PlayerAll.indexOf(otherCollider.group) >= 0) {
+            var player = otherCollider.getComponent(Player);
+            if (!GameConfig.invincibleMode && !player.invincible) {
+                player.isDying = true;
+            }
+        }
+    }
+    
     setSpeed(speedX: number, speedY: number) {
         this.speedX = speedX;
         this.speedY = speedY;
